moving existing unit tests to live in same package as tested class
